fixes for pr
